fix revalidation of cases when cases are settled
